+++ frontend/src/components/footer.tsx
@@ -1,7 +1,9 @@
+const APP_VERSION = __APP_VERSION__;
+
 export default function Footer() {
   return (
     <div className="container">
-      IRR explorer v2.4.0 ·{' '}
+      IRR explorer v{APP_VERSION} ·{' '}
       <a className="link-dark" href="https://www.nlnog.net/">
         Stichting NLNOG
       </a>{' '}
@@ -10,8 +12,8 @@ export default function Footer() {
           stichting@nlnog.net
         </a>
       ) · Source on{' '}
-      <a className="link-dark" href="https://github.com/jskoetsier/irrexplorer">
-        GitHub
+      <a className="link-dark" href="https://gitlab.int.koetsier.org/sebas/irrexplorer">
+        GitLab
       </a>{' '}
       ·{' '}
       <a className="link-dark" href="/api/docs">